+++ xeHentai/i18n/zh_hant.py
@@ -60,7 +60,8 @@ XEH_OPT_a = "下載完成後生成zip壓縮包並刪除下載目錄 (當前: %(d
 XEH_OPT_j = "使用日語標題, 如果關閉則使用英文或羅馬字標題 (當前: %(default)s)"
 XEH_OPT_download_range = "設置下載的圖片範圍, 格式為 開始位置-結束位置, 或者單張圖片的位置, " \
 "使用逗號來分隔多個範圍, 例如 5-10,15,20-25, 默認為下載所有"
-XEH_OPT_timeout = "設置下載圖片的超時 (當前: %(default)ss)"
+XEH_OPT_timeout = "設置下載圖片的超時 (當前: %(default)s秒)"
+XEH_OPT_f = "忽略配額判斷，繼續下載 (當前: %(current)s)"
 XEH_OPT_h = "顯示本幫助信息"
 XEH_OPT_version = "顯示版本信息"
 XEH_OPT_IGNORING_I = "後台模式已忽略 -i 參數"